+++ packages/dangoui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # dangoui
 
+## 3.3.1
+
+### Patch Changes
+
+- 4f83529: fix: avatar reverse
+
+## 3.3.0
+
+### Minor Changes
+
+- 122b1de: feat: avatar
+
+## 3.2.0
+
+### Minor Changes
+
+- d0fa4d0: fix: miss found cascader default value bug
+
 ## 3.1.56
 
 ### Patch Changes